Fix #82 카운터 모듈에서 날짜 선택 기능을 이용한 XSS 보안 취약점 문제 고침 (XEVE-16-004)

+++ modules/counter/counter.admin.view.php
@@ -24,7 +24,7 @@
 		 **/
 		function dispCounterAdminIndex() {
 			// set today's if no date is given
-			$selected_date = Context::get('selected_date');
+			$selected_date = (int) Context::get('selected_date');
 			if(!$selected_date) $selected_date = date("Ymd");
 			Context::set('selected_date', $selected_date);
 			// create the counter model object
